Change: Move useUserName hook to hooks directory
