+++ src/web/components/menu/usermenu.jsx
@@ -23,7 +23,7 @@ import Link from 'web/components/link/link';
 
 import Theme from 'web/utils/theme';
 import useGmp from 'web/utils/useGmp';
-import useUserName from 'web/utils/useUserName';
+import useUserName from 'web/hooks/useUserName';
 import useUserSessionTimeout from 'web/utils/useUserSessionTimeout';
 import useUserTimezone from 'web/utils/useUserTimezone';
